+++ test/fixtures/meta_status_decorations/_targets.R
@@ -2,5 +2,6 @@ list(
 	tar_target(warn_only, 1),
 	tar_target(error_only, 2),
 	tar_target(clean_target, 3),
-	tar_target(not_built, 4)
+	tar_target(not_built, 4),
+	tar_target(canceled_target, 5)
 )